+++ tests/unit-testing/camera_tests.cpp
@@ -174,7 +174,7 @@ SPECTRE_TEST(Camera, Perspective_create_ray)
     Vec3 up(0.f, 1.f, 0.f);
 
     //square image
-    CameraPerspective camera(pos, target, up, 2, 2, radians(60));
+    CameraPerspective camera(pos, target, up, 2, 2, radians((float)60));
     Sample samples[4];
     sam.get_samples(samples);
     Ray r0, r1, r2, r3;
@@ -215,7 +215,7 @@ SPECTRE_TEST(Camera, Perspective_create_ray)
     EXPECT_NEAR(r3.direction.z, r3direction.z, 1e-5f);
 
     //horizontal image
-    camera = CameraPerspective(pos, target, up, 2, 1, radians(60));
+    camera = CameraPerspective(pos, target, up, 2, 1, radians((float)60));
     ((Camera*)&camera)->create_ray(samples, &r0);
     ((Camera*)&camera)->create_ray(samples+1, &r1);
     ((Camera*)&camera)->create_ray(samples+2, &r2);
@@ -267,7 +267,7 @@ SPECTRE_TEST(Camera, Perspective_create_ray_diff)
     Vec3 up(0.f, 1.f, 0.f);
 
     //square image
-    CameraPerspective camera(pos, target, up, 2, 2, radians(60));
+    CameraPerspective camera(pos, target, up, 2, 2, radians((float)60));
     Sample samples[4];
     sam.get_samples(samples);
